Refactor types and models
