+++ packages/phoenix-event-display/src/lib/models/cut.model.ts
