--- packages/phoenix-event-display/src/managers/three-manager/types/event-display.types.ts
@@ -1,6 +1,0 @@
-import { Object3D } from 'three';
-
-export type GeometryUIParameters = {
-  object: Object3D;
-  menuNodeName?: string;
-};
